SSHD-1055: Half-close port-forwarded socket on EOF in tunnel
Propagate an EOF on a tunnel for port forwarding to the socket by
shutting down its output; otherwise the client connected to that
socket may hang.

For the MINA transport back-end, calling socket.shutdownOutput(), which
is possible only via reflection, leads to problems as it confuses the
main polling loop inside MINA. So don't do that but schedule the MINA
session to close once all pending writes have been written.

Enable half-closure in the Netty transport back-end.

Fix the Nio2 transport back-end to not re-try accepting connections
when the channel is already closing.

See also SSHD-964 and SSHD-902.