SSHD-1269: Fix TCP/IP remote port forwarding with wildcard addresses
Remote port forwarding using an OpenSSH client and an Apache MINA sshd
server didn't work with wildcard addresses or with "localhost". For
instance,

ssh ... -R 0.0.0.0:0:somewhere:1234

would fail: the Apache MINA sshd server would send a forwarded-tcpip
request with 127.0.0.1:55555 (if port 55555 was chosen) to OpenSSH,
but OpenSSH expected 0.0.0.0:55555.

ssh ... -R 0:somewhere:1234

also failed in the same way.

Fix this by sending back in the forwarded-tcpip request the original
address with the bound port.

(Note: this was already fixed by SSHD-792 as of Apache MINA sshd 2.2.0,
but was broken again in 2.6.0. Unit tests using an Apache MINA sshd
client to initiate the remote port forwarding didn't detect the problem
because Apache MINA sshd only checks the port but not the hostname.)