[SSHD-822] Netty fix for ClientDeadlockTest
This test failed on Netty because Netty swallows and logs exceptions
propagated out of exceptionCaught(). The server in this test was left
with a NettyIoSession without SSH session.

The fix is to close the NettyIoSession if we get an exception inside
channelActivated() when trying to create the SSH session for it.

Also remove the ChannelOption.ALLOW_HALF_CLOSURE in NettyIoSession.
It was added in commit d2f0b13e, but it's actually wrong. It prevents
closing a channel when the peer closes its output. It is not needed
if we want to shut down *our* output, which is what commit d2f0b13e was
about. With the option set, the ClientDeadlockTest would sometimes fail
because the client would not close the ClientSession when the server
closed the channel.