[SSHD-822] Enable ApacheServer*Test for MINA
Enable the following two tests for MINA:

* ApacheServerApacheClientTest
* ApacheServerJSchClientTest

The classes contain fundamentally flawed tests though that assume that
large-ish data can be read in a single read() call (or in two read()
calls).

This can only work if all buffers in all layers are large enough to
hold the data. With MINA, this is not the case: MINA manages the read
buffer size for a channel/session *dynamically*, and it starts out at
2048 bytes. Therefore these tests *cannot* work with MINA, and are
disabled for MINA.

The other tests in these classes are enabled for MINA.