SCP: Call the receive listener before sending back OK
The listener may interdict the operation by throwing an exception. If
we've already sent an OK, it's possible that the client goes ahead and
tries to write the file data, and then gets an exception because the
channel is closed.

By calling the listener earlier we ensure that we send back a NACK, and
the client won't even try to upload anything.