[SSHD-1293] Fix unbinding port forwarding for auto-alloc port
Fix unbinding a local TCP/IP port forwarding binding when using a
dynamic auto-allocated port. "localAddress" is controlled by the
user and may contain still a port 0, which means auto-allocating
a local port; the actually bound local port is available in
"boundAddress".

The DefaultForwarder tracks the actual bound port mappings only,
so "boundAddress" must be used to unbind the port.