[SSHD-1295] Close sessions when ConnectFuture.verify() times out
Connections are established asynchronously: a connection attempt
continues even if ConnectFuture.verify(timeout) times out. If the
connection then succeeds, there is a connected ClientSession that
the user doesn't know about and cannot use or close.

Detect this case in SshClient and close the sessions if there was
already a timeout on the future.