[SSHD-822] Fix MINA transport for load tests and enable them
The load tests do run fine with Apache MINA. But enabling them
showed abysmal performance in port forwarding. Transferring about
400kB locally took about 10 seconds!

The reason is a possible bug in Apache MINA triggered by the recent
re-write of TCP/IP port forwarding. Commit b91a4245 changed the port
forwarding such that it disabled reads on the I/O port session
(IoSession.suspendRead()) while writing to the SSH tunnel, and re-
enabling reads (IoSession.resumeRead()) once the write was complete.
This avoids the need to buffer data inside Apache MINA; it throttles
reading from the port to the speed of writing through the tunnel.

This works fine with the NIO2 and Netty transports, but with Apache
MINA re-enabling reads becomes effective only on the next select call.
If a select call was already on-going, re-enabling reading might become
effective only after one second (the MINA idle time-out on the select).

Fix this by forcing an on-going select call to return immediately when
reading on a MINA IoSession is re-enabled. Apache MINA thus picks up
the change immediately and starts a new select() with the change right
away.