TCP/IP port forwarding: fix closing ChannelAsyncOutputStream
The Channels using a ChannelAsyncOutputStream  must allow writing while
being closed, otherwise the ChannelAsyncOutputStream cannot be closed
gracefully because it uses the channel as its packet writer. When the
Channel is closed, it closes its ChannelAsyncOutputStream first, but by
that time the channel state is already "closing" and writes are normally
no longer allowed.