[SSHD-1295] Handle cancellation and time-out on futures
Factor out cancellation operations and exception reporting on futures
into separate interfaces. Mark futures as canceled by storing a
CancelFuture. A separate future is needed because not all cancellations
are effective immediately. A CancelFuture may have a backtrace giving
an indication of where the future was cancelled.

In AbstractSshFuture.verifyResult(), give the SshException on time-out
a TimeoutException as cause so that clients can recognize time-outs
without having to resort to parsing an exception message. Also handle
the case of the future having been fulfilled with a null marker, or
having been cancelled.

Give all verify() and await() methods optional CancelOption parameters.
There are three CancelOptions: CANCEL_ON_TIMEOUT, CANCEL_ON_INTERRUPT,
and NO_CANCELLATION. By passing the appropriate options, clients can
control whether they want the future and thus the operation it describes
to be cancelled.

For the DefaultConnectFuture, implement verify() such that by default
it uses CANCEL_ON_TIMEOUT and CANCEL_ON_INTERRUPT. This covers the
normal use case in existing code, where there is exactly one thread
calling verify(). In this scenario, not cancelling would lead to the
connection attempt to continue and possibly succeed, at which point
there'd be a connection, but the client code has no way to get at the
created session in any reasonable way to close it.

With previous code, it was possible that different threads might call
verify() on the same future with different time-outs. In such cases,
client code should be adapted as needed; at least such that all but one
thread use NO_CANCELLATION, and the final thread (probably the one using
the longest time-out) just calling verify() or using CANCEL_ON_TIMEOUT
and CANCEL_ON_INTERRUPT explicitly.

Besides ConnectFuture.verify(timeout) the underlying I/O back-end can
also have its own connection timeout. Give users a new property
CoreModuleProperties.IO_CONNECTION_TIMEOUT to control the I/O library's
connection time-out.

When a future is canceled, whether via cancel() or due to a time-out or
interrupt, pass down the cancellation to any lower-level futures via
listeners if the appropriate CancelOption is set. This has the effect
that if client code cancels a future, dependent futures that are
invisible to the client are also cancelled. For a DefaultConnectFuture
the cancellation thus propagates to the I/O connection future of the
transport back-end, which ensures that all Sessions, IoSessions, and
even low-level connections are properly closed, canceled, or aborted.

Implement cancellation for user authentication. The authentication
protocol can be canceled only if one does not possibly expect a success
reply, and any cancellation can take effect only once an outstanding
reply in the protocol has been received.

For the OpenFuture, handle cancellation by closing the channel right
away again if opened successfully. It appears, though, that some
channels cannot be canceled, simply because Apache MINA sshd presents
the user with a synchronous interface. An example are TCP/IP port
forwardings.