[SSHD-1173] Simplify writing in ChannelAsyncOutputStream
Remove the option not to send a chunk when the remote window is
smaller than the packet size, and we're trying to write more than
the remote window size bytes. This option may lead to hangs with
peers that wait with sending their window adjustment until the window
size is very low, or even zero.

RFC 4254 does not mandate any particular point at which an SSH
implementation should send a window adjustment. Implementations
doing so only once the window has been fully consumed are well
within the bounds of the specification.

Therefore remove this option altogether, and always send some data
if the window size is > 0. Refactor the writePacket() method to make
it much simpler and clearer.

Client options removed:

* SftpModuleProperties.CHUNK_IF_WINDOW_LESS_THAN_PACKET
* CoreModuleProperties.ASYNC_SERVER_STDOUT_CHUNK_BELOW_WINDOW_SIZE
* CoreModuleProperties.ASYNC_SERVER_STDERR_CHUNK_BELOW_WINDOW_SIZE
* ChannelAsyncOutputStream(Channel, byte, boolean) constructor

These options were by default all false, potentially leading to hangs
as described above. Already [SSHD-1123] had pointed out that problem,
and bug [SSHD-1207] might also be caused by this. The feature was
introduced in [SSHD-979] in commit e85b67e0 and later modified in
commit 536d0663.

Also fix a race condition in WindowAdjustTest (unsynchronized access)
that made the test frequently hang because the write queue maintained
in that test got corrupted.