[releng] Repository configuration for GitHub
Per [1] the ASF JIRA requires manual account creation as of 2022-11-06.
This manual workflow is not manageable for us and introduces a new and
deterring barrier for new prospective issue reporters. Therefore enable
issue management via GitHub issues.

People with existing JIRA accounts may continue to use the ASF JIRA.

Add a minimal ASF configuration[2] for the GitHub repository; tell in
the README where to report issues. Also add the clear-text security
e-mail in security.txt: it makes no sense to have our mailing lists in
clear in the repo, but keep the security e-mail "hidden".

[1] https://lists.apache.org/thread/jx9d7sp690ro660pjpttwtg209w3m39w
[2] https://s.apache.org/asfyaml