TcpipClientChannel: update local window only after data was written
Like the TcpipServerChannel, the TcpipClientChannel must update the
local channel window only once the received data has actually been
forwarded to the socket.

Factor out the code for forwarding from an SSH channel to a port,
and use the same code in both the TcpipClientChannel and in the
TcpipServerChannel. Forwarding data from the SSH channel to the
socket should be identical on both sides of the tunnel.