[releng] Update GitHub action versions
Use @v3 versions; the old versions are no longer supported and will
stop working in a few months.

Use the Temurin JDK distro. Temurin doesn't have Java 14, and anyway
that was an intermediary version. Use the Java 17 LTS instead.

Note that [DIRSTUDIO-1277] forces us to use --add-opens for the
LDAP tests on Java >= 16.[1]

[1] https://issues.apache.org/jira/browse/DIRSTUDIO-1277