GH-268: Fix ClientConnectionService.sendHeartBeat()
If a reply is requested, but none arrives within the timeout, throw an
exception and terminate the connection.

This rolls back the changes made in ClientConnectionService in commit
de2f8fef. It's quite all right to use the synchronous implementation of
Session.request() because heartbeats are not sent from I/O threads.

However, commit de2f8fef broke the contract specified in interface
Session, which says Session.request() must return "the buffer if the
request was successful, {@code null} otherwise." The implementation
from commit de2f8fef threw an exception instead.

This was wrong, and is corrected now in this commit. Note that this
means that a caller cannot distinguish between the server replying
SSH_MSG_UNIMPLEMENTED or SSH_MSG_REQUEST_FAILURE.

Where such distinction is needed, use an asynchronous global request
instead.

Bug: https://github.com/apache/mina-sshd/issues/268