GH-282: Fix ACL for host key file on Windows
Apparently it is possible on Windows that the ACL for a file have no
entry for the file owner. (Typically the owner is then member of a
group, and there is an entry for the group.) The previous code would in
such cases remove all access permissions from all entries, resulting in
a file that even the owner could no longer access.

Change the logic: not only remove all permissions for non-owner entries
but also add an explicit entry for the owner, giving it full access
permissions.

Bug: https://github.com/apache/mina-sshd/issues/282