GH-285: Fix compilation failure on Java 19
ExecutorService implements AutoCloseable in Java 19 and provides a
default method for close(). This conflicts in CloseableExecutorService,
which mixes in org.apache.sshd.common.Closeable, which has its own
default method for close().

As a minimum fix add an overrides of close() explicitly calling
the Apache MINA sshd variant of close().

Bug: https://github.com/apache/mina-sshd/issues/285