GH-298: Start service in AbstractServerSession
Services need to be started. In particular, the connection service
starts its heartbeat, if configured, when started. It was already
noticed in commit e74e2041 that the server-side did not start services,
but at that time the effect was not recognized because tests were too
unstable back then. A test for server-side heartbeats was also missing.

Bug: https://github.com/apache/mina-sshd/issues/298