GH-283: Fix handling of CoreModuleProperties.PASSWORD_PROMPTS
The setting was not considered in password authentication, and in
keyboard-interactive authentication counted the number of non-empty
challenges, which could lead to an extra authentication request
message to be sent, or to keyboard-interactive authentication being
aborted too early.

The property is supposed to correspond to the OpenSSH configuration
NumberOfPasswordPrompts, which despite the name counts not the number
of prompts but the number of times password or keyboard-interactive
authentication is attempted. For password authentication, this is
the same (if one ignores possible password-change requests), but in
keyboard-interactive authentication, there may be several prompts
per attempt.

Fix the implementation in Apache MINA sshd to respect the limit in
both password and keyboard-interactive authentication, and in both
apply the limit to the count of interactive authentication attempts.

If password identities are provided on a session non-interactively
(via addPasswordIdentity(), or via a custom PasswordIdentityProvider
set with setPasswordIdentityProvider()), any authentication attempts
that are answered non-interactively do not count towards the limit.

Bug: https://github.com/apache/mina-sshd/issues/283