GH-325: SFTP: fix deleting remote symbolic links
Files.delete() is specified *not* to follow symbolic links. The
implementation in SftpFileSystemProvider checked for write access
and file existence via a method that does resolve symbolic links.
This lead to general inconsistency and various kinds of failures.

Fix this by making sure that the code does not resolve symbolic
links in Files.delete().

Bug: https://github.com/apache/mina-sshd/issues/325