[releng] OSGi: Ensure the BC security provider can be found
BouncyCastleSecurityProviderRegistrar references the BC security
provider only reflectively. It lives in a package that is not
referenced explicitly anywhere.

To be able to find org.bouncycastle.jce.provider.BouncyCastleProvider,
the package must be on the bundle classpath in OSGi. It wasn't, since
there was no "normal" reference to it.

Add an optional dependency explicitly in the generated MANIFEST.MF of
sshd-osgi.