GH-364: Use flags for RSA signature requests in AbstractAgentClient
Respect the flags in OpenSSH signing requests. These flags tell what
kind of signature should be produced for an RSA key: ssh-rsa,
rsa-sha2-256, or rsa-sha2-512.

Previous code only set the flags on the client side, but ignored
them in the server side and thus always answered with an ssh-rsa
signature.

Bug: https://github.com/apache/mina-sshd/issues/364