GH-370: Also compare file keys in ModifiableFileWatcher
Additionally, handle the case of files being modified very quickly,
such that the last modified timestamp doesn't change, even though the
file was modified. If the modification did not change the file size,
such metadata is "racily clean", meaning it indicates the file had
not been modified when in fact it was.

This can occur because of the finite resolution of file timestamps.
If the file timestamp has a granularity of 2 seconds (FAT), two file
modifications within these two seconds that don't change the file
size cannot be recognized.

Hence any file timestamp from the past 2 seconds (measured from the time
it was read) is suspect, and the file must be considered potentially
modified, and must be re-loaded.

This is not a problem unless one frequently writes files and then reads
them again within two seconds. (Or if one reads the same file multiple
times within two seconds.) In such cases, care should be taken to
determine the actual resolution of file timestamps, which often is
much better. But for the use cases in SSH the worst-case assumption
of two seconds should be fine.

Bug: https://github.com/apache/mina-sshd/issues/370