[SSHD-1310] SftpFileSystem: do not close user sessions
A SftpFileSystem can be obtained in two ways:

1. Via FileSystems.newFileSystem()
2. Via SftpClientFactory.instance().createSftpFileSystem(ClientSession)

In the first case, the SftpFileSystemProvider automatically creates a
ClientSession, and that session must be closed when the file system
closes.

In the second case, user code already has a ClientSession and wants to
run an SftpFileSystem within. In this case, the session _must not_ be
closed when the file system closes.

Add a session attribute that can be set on a ClientSession to tell the
SftpFileSystem whether it owns the ClientSession. The SftpFileSystem
will close only owned sessions.

Also add a SessionListener to ensure the file system is closed when the
session closes.

Bug: https://issues.apache.org/jira/browse/SSHD-1310