SftpFileSystem: keep a session until the filesystem is closed
When an SftpFileSystem is obtained via FileSystems.newFileSystem(URI),
the new file system is created with an implicit SshClient and a new
implicit ClientSession. This session has a heartbeat set to keep it
open for the lifetime of the file system, i.e., until the file system
is closed.

But heartbeats do not guarantee that the session remains open. A server
may decide all the same to close the session, for instance, if it is
idle for too long. (Even despite heartbeat messages.)

So ensure that in such a case, the file system automatically tries to
open a new session. Otherwise a client of such an SftpFileSystem must
be prepared to deal with these cases at any time, which may be difficult
or even impossible to do.