GH-356: Publish snapshot artifacts in Github builds
On pushes to the master branch, publish snapshot artifacts to the
Apache snapshot repo if the tests were successful.

Do nothing for release versions; releases are prepared manually and
go through the normal Apache process (staging, voting) before being
released.

Snapshots are not and do not have to be signed according to [1].

The deployment jobs are serialized, and they deploy only if

* the project is indeed a -SNAPSHOT version
* the current master branch has not advanced
* the commit message does not start with "[maven-release-plugin]"

[1] https://issues.apache.org/jira/browse/INFRA-24609

Bug: https://github.com/apache/mina-sshd/issues/356