GH-388: ChannelExec: support non-UTF-8 commands
Add operations to execute commands that use some other encoding than
UTF-8: one version that has a Charset as extra parameter, and another
version that simply takes a byte[] as command.

Use the new way to create a ChannelExec in the SCP client to encode the
command using the ScpModuleProperties.SCP_OUTGOING_ENCODING.

Add an SCP test against an OpenSSH daemon running in a container using
ISO-8859-15 as locale.

Bug: https://github.com/apache/mina-sshd/issues/388