GH-403: Work-around for WS_FTP client bug
If the file handle size in Apache MINA sshd is > 4, WS_FTP client <=
12.9 sends (fileHandleSize - 4) too many bytes in SSH_FXP_WRITE
requests. If that exceeds the maximum packet size of the channel's
local window, the channel packet is invalid and Apache MINA sshd
correctly throws an exception and closes the channel.

Work around this by allowing for this special case on the server side
in channels used for SFTP.

Bug: https://github.com/apache/mina-sshd/issues/403